requirements.txt
streamlit>=1.33
pandas>=2.2
gspread>=6.0
google-auth>=2.31
xlsxwriter>=3.2
